fast_monitor/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiochclient==2.0.0
    # via -r requirements.in
aiographql-client==1.0.2
    # via -r requirements.in
aiohttp==3.7.4.post0
    # via
    #   -r requirements.in
    #   aiochclient
    #   aiographql-client
asgiref==3.3.4
    # via uvicorn
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via aiohttp
cafeteria==0.20.0
    # via cafeteria-asyncio
cafeteria-asyncio==0.2.1
    # via aiographql-client
certifi==2021.5.30
    # via
    #   httpx
    #   sentry-sdk
chardet==4.0.0
    # via aiohttp
click==8.0.1
    # via uvicorn
clickhouse-driver==0.2.1
    # via -r requirements.in
cycler==0.10.0
    # via matplotlib
fastapi==0.65.2
    # via -r requirements.in
fluent==0.10.0
    # via -r requirements.in
fluent-logger==0.10.0
    # via -r requirements.in
graphql-core==3.1.5
    # via aiographql-client
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.12.3
    # via httpx
httpx==0.16.1
    # via -r requirements.in
idna==2.10
    # via
    #   rfc3986
    #   yarl
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.4.2
    # via -r requirements.in
msgpack==1.0.2
    # via fluent-logger
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
numpy==1.21.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
pandas==1.3.0
    # via -r requirements.in
pillow==8.2.0
    # via
    #   -r requirements.in
    #   matplotlib
pydantic[dotenv]==1.8.2
    # via
    #   -r requirements.in
    #   fastapi
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via
    #   matplotlib
    #   pandas
python-dotenv==0.18.0
    # via pydantic
pytz==2021.1
    # via
    #   clickhouse-driver
    #   pandas
    #   tzlocal
pyyaml==5.4.1
    # via
    #   -r requirements.in
    #   cafeteria
rfc3986[idna2008]==1.5.0
    # via httpx
sentry-sdk==1.1.0
    # via -r requirements.in
six==1.16.0
    # via
    #   cycler
    #   python-dateutil
sniffio==1.2.0
    # via
    #   httpcore
    #   httpx
sqlparse==0.4.1
    # via aiochclient
starlette==0.14.2
    # via fastapi
typing-extensions==3.10.0.0
    # via
    #   aiohttp
    #   pydantic
tzlocal==2.1
    # via clickhouse-driver
ujson==3.2.0
    # via aiographql-client
urllib3==1.26.6
    # via sentry-sdk
uvicorn==0.14.0
    # via -r requirements.in
yarl==1.6.3
    # via aiohttp
